requirements.txt
asgiref==3.4.1
blinker==1.4
brotli==1.0.9
contextvars==2.4
cryptography==3.4.7
docutils==0.17.1
flask_httpauth==4.4.0
flask_sqlalchemy==2.5.1
importlib_metadata==4.6.1
keyring==23.0.1
monotonic==1.6
numpy==1.21.0
ordereddict==1.1
passlib==1.7.4
protobuf==3.17.3
python-dotenv==0.18.0
tornado==6.1
typing_extensions==3.10.0.0
watchdog==2.1.3
psycopg2==2.9.1
gunicorn==20.1.0
zipp==3.5.0
eventlet==0.33.0
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
